requirements.txt
asgiref==3.11.1
pdf2image==1.17.0
dj-database-url==3.1.2
Django==5.2.13
gunicorn==25.3.0
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
whitenoise==6.12.0
google-cloud-documentai>=2.29.0
google-cloud-vision>=3.7.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
PyMuPDF>=1.23.0
